requirements.txt
asgiref==3.6.0
crispy-bootstrap5==0.7
Django==3.2.18
django-crispy-forms==2.0
django-filter==23.1
pytz==2023.3
sqlparse==0.4.3
typing-extensions==4.5.0
